+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/service/impl/runner/AuditRequestConsumerServiceImpl.java
@@ -61,7 +61,8 @@ public class AuditRequestConsumerServiceImpl extends AuditRequestServiceSyncStan
             MailService mailService,
             MessageService messageService,
             ActRepository actRepository,
-            ContractUserRepository contractUserRepository) {
+            ContractUserRepository contractUserRepository,
+            ProjectUserRepository projectUserRepository) {
 
         super(pageRepository,
                 auditRepository,
@@ -76,7 +77,8 @@ public class AuditRequestConsumerServiceImpl extends AuditRequestServiceSyncStan
                 mailService,
                 messageService,
                 actRepository,
-                contractUserRepository);
+                contractUserRepository,
+                projectUserRepository);
         this.auditRequestConsumer = auditRequestConsumer;
     }